requirements/base.txt
django==5.0.7
django-jazzmin==3.0.0
django-modeltranslation==0.19.4
django-rosetta==0.10.0

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

python-environ==0.4.54
python-telegram-bot==13.15

# Database
psycopg2-binary==2.9.9

drf-yasg==1.21.7

setuptools==71.0.3

django-cors-headers==4.4.0

Pillow==10.4.0

langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
langchainhub==0.1.21

faiss-cpu
